requirements.txt
flask==3.0.2
pandas==2.2.0
pytz==2024.1
requests~=2.32.3
python-dotenv~=1.1.0